Refactor commands package outside of gbb package
