+++ commands/command_test.go
