+++ commands/download_test.go
